Unnest nested lists

Unnest spec is a nested list with the same structure as the nested json. It specifies concisely how the deeply nested components ought to be unnested. s() is a shorthand for spec() and returns the canonical spec (a spec which can be consumed directly by the C++ routines).

s(
  selector = NULL,
  ...,
  as = NULL,
  children = NULL,
  groups = NULL,
  include = NULL,
  exclude = NULL,
  stack = NULL,
  process = NULL
)

spec(
  selector = NULL,
  ...,
  as = NULL,
  children = NULL,
  groups = NULL,
  include = NULL,
  exclude = NULL,
  stack = NULL,
  process = NULL
)

unnest(
  x,
  spec = NULL,
  dedupe = FALSE,
  stack_atomic = FALSE,
  process_atomic = NULL,
  cross_join = TRUE
)

Arguments

selector
A shorthand syntax for an include selector. When a list, each element of the list is expanded into the include element at the respective level. When a string, it is expanded into a list according to the following rules:

When selector is of length 1 and contains "/" characters it is split with the "/" separator. For instance s(c("a", "b"), ...), s("a/b", ...) and s("a", s("b", ...)) are all converted to the canonical s(include = "a", s(include = "b", ...)). Components consisting entirely of digits are converted to integers. For example s("a/2/b" ...) is equivalent to s("a", s(2, s("b", ...)))

Each element of the resulting from the previous step vector is split with ,. Thus s("a/b,c/d") is equivalent to s("a", s(include = c("b", "c"), s("d", ...)))

as
name for this field in the extracted data.frame

children, ...
Unnamed list of children spec. ... is merged into children. children is part of the canonical spec.

groups
Named list of specs to be processed in parallel. The return value is a named list of unnested data.frames. The results is the same as when each spec is unnested separately except that dedupe parameter of unnest() will work across groups and execution is faster because the nested list is traversed once regardless of the number of groups.

include, exclude
A list, a numeric vector or a character vector specifying components to include or exclude. A list can combine numeric indexes and character elements to extract.

stack
Whether to stack this node (TRUE) or to spread it (FALSE). When stack is a string an index column is created with that name.

process
Extra processing step for this element. Either NULL for no processing (the default), "asis" to return the entire element "as is" in a list column, "paste" to paste elements together into a character column.

x
a nested list to unnest

spec
spec to use for unnesting. See spec().

dedupe
whether to dedupe repeated elements. If TRUE, if a node is visited for a second time and is not explicitly declared in the spec the node is skipped. This is particularly useful with grouped specs.

stack_atomic
Whether atomic leaf vectors should be stacked or not.

process_atomic
Process spec for atomic leaf vectors. Unstable: might be removed in the future as it doesn't seem to be useful

cross_join
Specifies how the results from sibling nodes are joined (cbind) together. The shorter data.frames (in terms o number of rows) can be either recycled to the max number of rows across all components as with standard R's recycling (cross_join = FALSE). Or, with cross_join = TRUE, the results are cross joined (aka form all combinations of rows across joined components). cross_join = TRUE is the default because of no data loss and it is more conducive for earlier error detection with incorrect specs.

Value

spec(): a canonical spec - a list consumed by C++ unnesting routines.

Examples

## `s()` returns a canonical spec list
s("a")

#> <unnest.spec>
#>  $ include:"a"

s("a//c2")

#> <unnest.spec>
#>  $ include :"a"
#>  $ children:List of 1
#>   ..$ :<unnest.spec>
#>   .. ..$ children:List of 1
#>   .. .. ..$ :<unnest.spec>
#>   .. .. .. ..$ include:"c2"

s("a/2/c2,cid")

#> <unnest.spec>
#>  $ include :"a"
#>  $ children:List of 1
#>   ..$ :<unnest.spec>
#>   .. ..$ include :2
#>   .. ..$ children:List of 1
#>   .. .. ..$ :<unnest.spec>
#>   .. .. .. ..$ include:"c2" "cid"

x <- list(a = list(b = list(x = 1, y = 1:2, z = 10),
                   c = list(x = 2, y = 100:102)))
xxx <- list(x, x, x)

## spreading
unnest(x, s("a"))

#>   a.b.x a.b.y a.b.y.2 a.b.z a.c.x a.c.y a.c.y.2 a.c.y.3
#> 1     1     1       2    10     2   100     101     102

unnest(x, s("a"), stack_atomic = TRUE)

#>   a.b.x a.b.y a.b.z a.c.x a.c.y
#> 1     1     1    10     2   100
#> 2     1     2    10     2   101
#> 3     1     1    10     2   102
#> 4     1     2    10     2   100
#> 5     1     1    10     2   101
#> 6     1     2    10     2   102

unnest(x, s("a/b"), stack_atomic = TRUE)

#>   a.b.x a.b.y a.b.z
#> 1     1     1    10
#> 2     1     2    10

unnest(x, s("a/c"), stack_atomic = TRUE)

#>   a.c.x a.c.y
#> 1     2   100
#> 2     2   101
#> 3     2   102

unnest(x, s("a"), stack_atomic = TRUE, cross_join = TRUE)

#>   a.b.x a.b.y a.b.z a.c.x a.c.y
#> 1     1     1    10     2   100
#> 2     1     2    10     2   101
#> 3     1     1    10     2   102
#> 4     1     2    10     2   100
#> 5     1     1    10     2   101
#> 6     1     2    10     2   102

unnest(x, s("a//x"))

#>   a.b.x a.c.x
#> 1     1     2

unnest(x, s("a//x,z"))

#>   a.b.x a.b.z a.c.x
#> 1     1    10     2

unnest(x, s("a/2/x,y"))

#>   a.c.x a.c.y a.c.y.2 a.c.y.3
#> 1     2   100     101     102

## stacking
unnest(x, s("a/", stack = TRUE))

#>   a.x a.y a.y.2 a.y.3 a.z
#> 1   1   1     2    NA  10
#> 2   2 100   101   102  NA

unnest(x, s("a/", stack = TRUE, as = "A"))

#>   A.x A.y A.y.2 A.y.3 A.z
#> 1   1   1     2    NA  10
#> 2   2 100   101   102  NA

unnest(x, s("a/", stack = TRUE, as = "A"), stack_atomic = TRUE)

#>   A.x A.y A.z
#> 1   1   1  10
#> 2   1   2  10
#> 3   2 100  NA
#> 4   2 101  NA
#> 5   2 102  NA

unnest(x, s("a/", stack = "id"), stack_atomic = TRUE)

#>   a.id a.x a.y a.z
#> 1    b   1   1  10
#> 2    b   1   2  10
#> 3    c   2 100  NA
#> 4    c   2 101  NA
#> 5    c   2 102  NA

unnest(x, s("a/", stack = "id", as = ""), stack_atomic = TRUE)

#>   id x   y  z
#> 1  b 1   1 10
#> 2  b 1   2 10
#> 3  c 2 100 NA
#> 4  c 2 101 NA
#> 5  c 2 102 NA

unnest(xxx, s(stack = "id"))

#>   a.b.x a.b.y a.b.y.2 a.b.z a.c.x a.c.y a.c.y.2 a.c.y.3 id
#> 1     1     1       2    10     2   100     101     102  1
#> 2     1     1       2    10     2   100     101     102  2
#> 3     1     1       2    10     2   100     101     102  3

unnest(xxx, s(stack = "id"), stack_atomic = TRUE)

#>    a.b.x a.b.y a.b.z a.c.x a.c.y id
#> 1      1     1    10     2   100  1
#> 2      1     2    10     2   101  1
#> 3      1     1    10     2   102  1
#> 4      1     2    10     2   100  1
#> 5      1     1    10     2   101  1
#> 6      1     2    10     2   102  1
#> 7      1     1    10     2   100  2
#> 8      1     2    10     2   101  2
#> 9      1     1    10     2   102  2
#> 10     1     2    10     2   100  2
#> 11     1     1    10     2   101  2
#> 12     1     2    10     2   102  2
#> 13     1     1    10     2   100  3
#> 14     1     2    10     2   101  3
#> 15     1     1    10     2   102  3
#> 16     1     2    10     2   100  3
#> 17     1     1    10     2   101  3
#> 18     1     2    10     2   102  3

unnest(xxx, s(stack = "id", s("a/b/y/", stack = TRUE)))

#>   a.b.y id
#> 1     1  1
#> 2     2  1
#> 3     1  2
#> 4     2  2
#> 5     1  3
#> 6     2  3

## exclusion
unnest(x, s("a/b/", exclude = "x"))

#>   a.b.y a.b.y.2 a.b.z
#> 1     1       2    10

## dedupe
unnest(x, s("a", s("b/y"), s("b")), stack_atomic = TRUE)

#>   a.b.x a.b.y a.b.y a.b.z
#> 1     1     1     1    10
#> 2     1     2     2    10
#> 3     1     1     1    10
#> 4     1     2     2    10

unnest(x, s("a", s("b/y"), s("b")), dedupe = TRUE, stack_atomic = TRUE)

#>   a.b.x a.b.y a.b.z
#> 1     1     1    10
#> 2     1     2    10

## grouping
unnest(xxx, stack_atomic = TRUE,
       s(stack = TRUE,
         groups = list(first = s("a/b/x,y"),
                       second = s("a/b"))))

#> $first
#>   a.b.x a.b.y
#> 1     1     1
#> 2     1     2
#> 3     1     1
#> 4     1     2
#> 5     1     1
#> 6     1     2
#> 
#> $second
#>   a.b.x a.b.y a.b.z
#> 1     1     1    10
#> 2     1     2    10
#> 3     1     1    10
#> 4     1     2    10
#> 5     1     1    10
#> 6     1     2    10
#> 

unnest(xxx, stack_atomic = TRUE, dedupe = TRUE,
       s(stack = TRUE,
         groups = list(first = s("a/b/x,y"),
                       second = s("a/b"))))

#> $first
#>   a.b.x a.b.y
#> 1     1     1
#> 2     1     2
#> 3     1     1
#> 4     1     2
#> 5     1     1
#> 6     1     2
#> 
#> $second
#>   a.b.z
#> 1    10
#> 2    10
#> 3    10
#> 

## processing asis
str(unnest(xxx, s(stack = "id",
                  s("a/b/y", process = "asis"),
                  s("a/c", process = "asis"))))

#> 'data.frame':	3 obs. of  3 variables:
#>  $ a.b.y:List of 3
#>   ..$ : int  1 2
#>   ..$ : int  1 2
#>   ..$ : int  1 2
#>  $ a.c  :List of 3
#>   ..$ :List of 2
#>   .. ..$ x: num 2
#>   .. ..$ y: int  100 101 102
#>   ..$ :List of 2
#>   .. ..$ x: num 2
#>   .. ..$ y: int  100 101 102
#>   ..$ :List of 2
#>   .. ..$ x: num 2
#>   .. ..$ y: int  100 101 102
#>  $ id   : int  1 2 3

str(unnest(xxx, s(stack = "id", s("a/b/", process = "asis"))))

#> 'data.frame':	3 obs. of  4 variables:
#>  $ a.b.x:List of 3
#>   ..$ : num 1
#>   ..$ : num 1
#>   ..$ : num 1
#>  $ a.b.y:List of 3
#>   ..$ : int  1 2
#>   ..$ : int  1 2
#>   ..$ : int  1 2
#>  $ a.b.z:List of 3
#>   ..$ : num 10
#>   ..$ : num 10
#>   ..$ : num 10
#>  $ id   : int  1 2 3

str(unnest(xxx, s(stack = "id", s("a/b", process = "asis"))))

#> 'data.frame':	3 obs. of  2 variables:
#>  $ a.b:List of 3
#>   ..$ :List of 3
#>   .. ..$ x: num 1
#>   .. ..$ y: int  1 2
#>   .. ..$ z: num 10
#>   ..$ :List of 3
#>   .. ..$ x: num 1
#>   .. ..$ y: int  1 2
#>   .. ..$ z: num 10
#>   ..$ :List of 3
#>   .. ..$ x: num 1
#>   .. ..$ y: int  1 2
#>   .. ..$ z: num 10
#>  $ id : int  1 2 3

## processing paste
str(unnest(x, s("a/b/y", process = "paste")))

#> 'data.frame':	1 obs. of  1 variable:
#>  $ a.b.y: chr "1,2"

str(unnest(xxx, s(stack = TRUE, s("a/b/", process = "paste"))))

#> 'data.frame':	3 obs. of  3 variables:
#>  $ a.b.x: chr  "1" "1" "1"
#>  $ a.b.y: chr  "1,2" "1,2" "1,2"
#>  $ a.b.z: chr  "10" "10" "10"

str(unnest(xxx, s(stack = TRUE, s("a/b", process = "paste"))))

#> 'data.frame':	3 obs. of  1 variable:
#>  $ a.b: chr  "1,1:2,10" "1,1:2,10" "1,1:2,10"